change design pattern for code
