+++ src/app/config/config.js
@@ -1,0 +1,4 @@
+export default {
+    defaultPath: '/index',
+    basename: '/',
+};
