+++ src/app/modules/TaskManager/Task-Component/task-Item.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
-import * as actions from "./../actions/index";
+import * as actions from "../task-manager-action";
 import Moment from "react-moment";
 
 function TaskItem(props) {
