+++ src/app/modules/TaskManager/Task-Manager-Reducer/task_reducer.js
@@ -1,4 +1,4 @@
-import { ACTION_TYPES } from "../constants/Constants";
+import { ACTION_TYPES } from "../../../config/Constants";
 
 var initialState = {
   id: "",
