+++ src/app/modules/TaskManager/task-manager-action.js
@@ -1,4 +1,4 @@
-import { ACTION_TYPES } from "./../constants/Constants";
+import { ACTION_TYPES } from "../../config/Constants";
 
 export const listAll = () => {
   return {
