+++ src/app/modules/common/home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const HomeComponent = () => {
+    return (
+        <div className="text-white">
+            Home page
+        </div>
+    );
+}
+
+export default HomeComponent;
