--- src/common/Validation.js
@@ -1,9 +1,0 @@
-import {MESSAGES} from './../constants/Constants'
-
-export function validationStringEmpty (
-    value
-) {
-    return {
-        
-    };
-}
