+++ src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import { createStore } from "redux";
-import myReducer from "./reducers/index";
+import myReducer from "./app/shared/reducer/index";
 import { Provider } from "react-redux";
 
 const store = createStore(
